:hammer: Restructure enable/disable commands under mod namespace
